Allow the testGenerator function to handle warnings.
Instead of only passing messages through, this now passes the full
result object into the `testGenerator` function.  This allows the
`testGenerator` function to decide how to handle warnings versus errors.

An additional argument was added to allow this change to be backwards
compatible, but we should likely remove the middle argument when we are
ready for a major version bump.